refactor: register attendance page
